# parse in parallel