--- a/src/Language/RawTerm/Rule/PostRawStmt.hs
+++ b/src/Language/RawTerm/Rule/PostRawStmt.hs
@@ -1,0 +1,23 @@
+module Language.RawTerm.Rule.PostRawStmt () where
+
+-- data BaseRawStmt name
+--   = RawStmtDefine
+--       C
+--       (RawStmtKind name)
+--       (RT.RawDef name)
+--   | RawStmtDefineData
+--       C
+--       Hint
+--       (name, C)
+--       (Maybe (RT.Args RT.RawTerm))
+--       (SE.Series (RawConsInfo name))
+--       Loc
+--   | RawStmtDefineResource
+--       C
+--       Hint
+--       (name, C)
+--       (C, RT.RawTerm)
+--       (C, RT.RawTerm)
+--       C
+--   | RawStmtNominal C Hint (SE.Series (RT.RawGeist name, Loc))
+--   | RawStmtForeign C (SE.Series RawForeignItem)