… src/Kernel/Parse/Move/Internal/Discern.hs
… src/Kernel/Parse/Move/Internal/Handle/NameMap.hs
… src/Kernel/Parse/Move/Interpret.hs
… src/Kernel/Parse/Move/Parse.hs
… src/Language/RawTerm/Rule/RawStmt.hs