```sh
# before (core @ 379af2e)
❯ hyperfine "neut clean && neut build test"
Benchmark 1: neut clean && neut build test
  Time (mean ± σ):     750.0 ms ±   8.4 ms    [User: 1485.2 ms, System: 592.3 ms]
  Range (min … max):   738.5 ms … 763.4 ms    10 runs

# after (core @ 379af2e)
❯ hyperfine "neut clean && neut build test"
Benchmark 1: neut clean && neut build test
  Time (mean ± σ):     618.6 ms ±  14.2 ms    [User: 1489.6 ms, System: 571.8 ms]
  Range (min … max):   593.4 ms … 641.9 ms    10 runs
```